backend/backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.17.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect2==0.8.0
daphne==4.1.0
Django==4.2.4
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
fonttools==4.60.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyphen==0.17.2
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
